+++ src/main/java/com/bingblue/TaobaoTools/pojo/ManyOrderDetailExample.java
@@ -204,52 +204,52 @@ public class ManyOrderDetailExample {
             return (Criteria) this;
         }
 
-        public Criteria andIdEqualTo(Integer value) {
+        public Criteria andIdEqualTo(Long value) {
             addCriterion("id =", value, "id");
             return (Criteria) this;
         }
 
-        public Criteria andIdNotEqualTo(Integer value) {
+        public Criteria andIdNotEqualTo(Long value) {
             addCriterion("id <>", value, "id");
             return (Criteria) this;
         }
 
-        public Criteria andIdGreaterThan(Integer value) {
+        public Criteria andIdGreaterThan(Long value) {
             addCriterion("id >", value, "id");
             return (Criteria) this;
         }
 
-        public Criteria andIdGreaterThanOrEqualTo(Integer value) {
+        public Criteria andIdGreaterThanOrEqualTo(Long value) {
             addCriterion("id >=", value, "id");
             return (Criteria) this;
         }
 
-        public Criteria andIdLessThan(Integer value) {
+        public Criteria andIdLessThan(Long value) {
             addCriterion("id <", value, "id");
             return (Criteria) this;
         }
 
-        public Criteria andIdLessThanOrEqualTo(Integer value) {
+        public Criteria andIdLessThanOrEqualTo(Long value) {
             addCriterion("id <=", value, "id");
             return (Criteria) this;
         }
 
-        public Criteria andIdIn(List<Integer> values) {
+        public Criteria andIdIn(List<Long> values) {
             addCriterion("id in", values, "id");
             return (Criteria) this;
         }
 
-        public Criteria andIdNotIn(List<Integer> values) {
+        public Criteria andIdNotIn(List<Long> values) {
             addCriterion("id not in", values, "id");
             return (Criteria) this;
         }
 
-        public Criteria andIdBetween(Integer value1, Integer value2) {
+        public Criteria andIdBetween(Long value1, Long value2) {
             addCriterion("id between", value1, value2, "id");
             return (Criteria) this;
         }
 
-        public Criteria andIdNotBetween(Integer value1, Integer value2) {
+        public Criteria andIdNotBetween(Long value1, Long value2) {
             addCriterion("id not between", value1, value2, "id");
             return (Criteria) this;
         }
@@ -264,52 +264,52 @@ public class ManyOrderDetailExample {
             return (Criteria) this;
         }
 
-        public Criteria andHeadIdEqualTo(Integer value) {
+        public Criteria andHeadIdEqualTo(Long value) {
             addCriterion("head_id =", value, "headId");
             return (Criteria) this;
         }
 
-        public Criteria andHeadIdNotEqualTo(Integer value) {
+        public Criteria andHeadIdNotEqualTo(Long value) {
             addCriterion("head_id <>", value, "headId");
             return (Criteria) this;
         }
 
-        public Criteria andHeadIdGreaterThan(Integer value) {
+        public Criteria andHeadIdGreaterThan(Long value) {
             addCriterion("head_id >", value, "headId");
             return (Criteria) this;
         }
 
-        public Criteria andHeadIdGreaterThanOrEqualTo(Integer value) {
+        public Criteria andHeadIdGreaterThanOrEqualTo(Long value) {
             addCriterion("head_id >=", value, "headId");
             return (Criteria) this;
         }
 
-        public Criteria andHeadIdLessThan(Integer value) {
+        public Criteria andHeadIdLessThan(Long value) {
             addCriterion("head_id <", value, "headId");
             return (Criteria) this;
         }
 
-        public Criteria andHeadIdLessThanOrEqualTo(Integer value) {
+        public Criteria andHeadIdLessThanOrEqualTo(Long value) {
             addCriterion("head_id <=", value, "headId");
             return (Criteria) this;
         }
 
-        public Criteria andHeadIdIn(List<Integer> values) {
+        public Criteria andHeadIdIn(List<Long> values) {
             addCriterion("head_id in", values, "headId");
             return (Criteria) this;
         }
 
-        public Criteria andHeadIdNotIn(List<Integer> values) {
+        public Criteria andHeadIdNotIn(List<Long> values) {
             addCriterion("head_id not in", values, "headId");
             return (Criteria) this;
         }
 
-        public Criteria andHeadIdBetween(Integer value1, Integer value2) {
+        public Criteria andHeadIdBetween(Long value1, Long value2) {
             addCriterion("head_id between", value1, value2, "headId");
             return (Criteria) this;
         }
 
-        public Criteria andHeadIdNotBetween(Integer value1, Integer value2) {
+        public Criteria andHeadIdNotBetween(Long value1, Long value2) {
             addCriterion("head_id not between", value1, value2, "headId");
             return (Criteria) this;
         }
